update sql moi nhat, update xoauser, css booking cua toan

+++ src/main/java/com/apt/project4/config/CustomUserDetailsService.java
@@ -28,7 +28,7 @@ public class CustomUserDetailsService implements UserDetailsService {
 
     // This method is used by JWTAuthenticationFilter
     @Transactional
-    public UserDetails loadUserById(Long id) {
+    public UserDetails loadUserById(int id) {
         User user = userRepository.findById(id).orElseThrow(
                 () -> new UsernameNotFoundException("User not found with id : " + id)
         );
